make sure we can dynamically generate directed graph relationships on demand (rather than at insertion time)
