+++ backend/pkg/database/sqlite_repository.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/dominikbraun/graph"
 	sourceModel "github.com/fastenhealth/fasten-sources/clients/models"
 	"github.com/fastenhealth/fastenhealth-onprem/backend/pkg"
 	"github.com/fastenhealth/fastenhealth-onprem/backend/pkg/config"
@@ -14,7 +13,6 @@ import (
 	"github.com/glebarez/sqlite"
 	"github.com/google/uuid"
 	"github.com/sirupsen/logrus"
-	"golang.org/x/exp/slices"
 	"gorm.io/datatypes"
 	"gorm.io/gorm"
 	"net/url"
@@ -234,7 +232,7 @@ func (sr *SqliteRepository) UpsertRawResource(ctx context.Context, sourceCredent
 
 	//create associations
 	//note: we create the association in the related_resources table **before** the model actually exists.
-
+	//note: these associations are not reciprocal, (i.e. if Procedure references Location, Location may not reference Procedure)
 	if rawResource.ReferencedResources != nil && len(rawResource.ReferencedResources) > 0 {
 		for _, referencedResource := range rawResource.ReferencedResources {
 			parts := strings.Split(referencedResource, "/")
@@ -250,7 +248,15 @@ func (sr *SqliteRepository) UpsertRawResource(ctx context.Context, sourceCredent
 				},
 				RelatedResourceFhir: nil,
 			}
-			err := sr.AddReciprocalResourceAssociations(ctx, &source, wrappedResourceModel, &source, relatedResource)
+			err := sr.AddResourceAssociation(
+				ctx,
+				&source,
+				wrappedResourceModel.SourceResourceType,
+				wrappedResourceModel.SourceResourceID,
+				&source,
+				relatedResource.SourceResourceType,
+				relatedResource.SourceResourceID,
+			)
 			if err != nil {
 				sr.Logger.Errorf("Error when creating a reciprocal association for %s: %v", referencedResource, err)
 			}
@@ -417,283 +423,31 @@ func (sr *SqliteRepository) GetPatientForSources(ctx context.Context) ([]models.
 	return wrappedResourceModels, results.Error
 }
 
-// Retrieve a list of all fhir resources (vertex), and a list of all associations (edge)
-// Generate a graph
-// return list of root nodes, and their flattened related resources.
-func (sr *SqliteRepository) GetFlattenedResourceGraph(ctx context.Context) ([]*models.ResourceFhir, []*models.ResourceFhir, error) {
-	currentUser, currentUserErr := sr.GetCurrentUser(ctx)
-	if currentUserErr != nil {
-		return nil, nil, currentUserErr
-	}
-
-	// Get list of all resources
-	wrappedResourceModels, err := sr.ListResources(ctx, models.ListResourceQueryOptions{})
-	if err != nil {
-		return nil, nil, err
-	}
-
-	// Get list of all relationships
-	var relatedResourceRelationships []models.RelatedResource
-
-	// SELECT * FROM related_resources WHERE user_id = "53c1e930-63af-46c9-b760-8e83cbc1abd9";
-	result := sr.GormClient.WithContext(ctx).
-		Table("related_resources").
-		Where(models.RelatedResource{
-			ResourceFhirUserID: currentUser.ID,
-		}).
-		Scan(&relatedResourceRelationships)
-	if result.Error != nil {
-		return nil, nil, result.Error
-	}
-
-	//Generate Graph
-	// TODO optimization: eventually cache the graph in a database/storage, and update when new resources are added.
-	g := graph.New(resourceVertexId, graph.Directed(), graph.Acyclic(), graph.Rooted())
-
-	//add vertices to the graph (must be done first)
-	for ndx, _ := range wrappedResourceModels {
-		err = g.AddVertex(
-			&wrappedResourceModels[ndx],
-		)
-		if err != nil {
-			return nil, nil, fmt.Errorf("an error occurred while adding vertex: %v", err)
-		}
-	}
-
-	//add edges to graph
-	for _, relationship := range relatedResourceRelationships {
-		err = g.AddEdge(
-			resourceKeysVertexId(relationship.ResourceFhirSourceID.String(), relationship.ResourceFhirSourceResourceType, relationship.ResourceFhirSourceResourceID),
-			resourceKeysVertexId(relationship.RelatedResourceFhirSourceID.String(), relationship.RelatedResourceFhirSourceResourceType, relationship.RelatedResourceFhirSourceResourceID),
-		)
-		if err != nil {
-			//this may occur because vertices may not exist
-			sr.Logger.Warnf("ignoring, an error occurred while adding edge: %v", err)
-		}
-	}
-
-	//// simplify graph if possible.
-	//graph.TransitiveReduction(g)
-
-	// AdjacencyMap computes and returns an adjacency map containing all vertices in the graph.
-	//
-	// There is an entry for each vertex, and each of those entries is another map whose keys are
-	// the hash values of the adjacent vertices. The value is an Edge instance that stores the
-	// source and target hash values (these are the same as the map keys) as well as edge metadata.
-	//	map[string]map[string]Edge[string]{
-	//		"A": map[string]Edge[string]{
-	//			"B": {Source: "A", Target: "B"}
-	//			"C": {Source: "A", Target: "C"}
-	//		}
-	//	}
-	adjacencyMap, err := g.AdjacencyMap()
-	if err != nil {
-		return nil, nil, fmt.Errorf("error while generating AdjacencyMap: %v", err)
-	}
-
-	// For a directed graph, PredecessorMap is the complement of AdjacencyMap. This is because in a directed graph, only
-	// vertices joined by an outgoing edge are considered adjacent to the current vertex, whereas
-	// predecessors are the vertices joined by an ingoing edge.
-	// ie. "empty" verticies in this map are "root" nodes.
-	predecessorMap, err := g.PredecessorMap()
-	if err != nil {
-		return nil, nil, fmt.Errorf("error while generating PredecessorMap: %v", err)
-	}
-
-	// Doing this in one massive function, because passing graph by reference is difficult due to generics.
-
-	// Step 1: use predecessorMap to find all "root" encounters and conditions. store those nodes in their respective lists.
-	encounterList := []*models.ResourceFhir{}
-	conditionList := []*models.ResourceFhir{}
-	for vertexId, val := range predecessorMap {
-
-		if len(val) != 0 {
-			//skip any nodes/verticies/resources that are not "root"
-			continue
-		}
-
-		resource, err := g.Vertex(vertexId)
-		if err != nil {
-			//could not find this vertex in graph, ignoring
-			continue
-		}
-
-		if strings.ToLower(resource.SourceResourceType) == "condition" || strings.ToLower(resource.SourceResourceType) == strings.ToLower(pkg.FhirResourceTypeComposition) {
-			conditionList = append(conditionList, resource)
-		} else if strings.ToLower(resource.SourceResourceType) == "encounter" {
-			encounterList = append(encounterList, resource)
-		}
-	}
-
-	// Step 2: define a function. When given a resource, should find all related resources, flatten the heirarchy and set the RelatedResourceFhir list
-	flattenRelatedResourcesFn := func(resource *models.ResourceFhir) {
-		// this is a "root" encounter, which is not related to any condition, we should add it to the Unknown encounters list
-		vertexId := resourceVertexId(resource)
-		sr.Logger.Debugf("populating resource: %s", vertexId)
-
-		resource.RelatedResourceFhir = []*models.ResourceFhir{}
-
-		//get all the resources associated with this node
-		graph.DFS(g, vertexId, func(relatedVertexId string) bool {
-			relatedResourceFhir, _ := g.Vertex(relatedVertexId)
-			//skip the current resource if it's referenced in this list.
-			//also skip the current resource if its a Binary resource (which is a special case)
-			if vertexId != resourceVertexId(relatedResourceFhir) && relatedResourceFhir.SourceResourceType != "Binary" {
-				resource.RelatedResourceFhir = append(resource.RelatedResourceFhir, relatedResourceFhir)
-			}
-			return false
-		})
-	}
-
-	// Step 3: populate related resources for each encounter, flattened
-	for ndx, _ := range encounterList {
-		// this is a "root" encounter, which is not related to any condition, we should add it to the Unknown encounters list
-		flattenRelatedResourcesFn(encounterList[ndx])
-
-		//sort all related resources (by date, desc)
-		encounterList[ndx].RelatedResourceFhir = utils.SortResourcePtrListByDate(encounterList[ndx].RelatedResourceFhir)
-
-	}
-
-	// Step 4: find all encounters referenced by the root conditions, populate them, then add them to the condition as RelatedResourceFhir
-	for ndx, _ := range conditionList {
-		// this is a "root" condition,
-
-		conditionList[ndx].RelatedResourceFhir = []*models.ResourceFhir{}
-		vertexId := resourceVertexId(conditionList[ndx])
-		for relatedVertexId, _ := range adjacencyMap[vertexId] {
-			relatedResourceFhir, _ := g.Vertex(relatedVertexId)
-			flattenRelatedResourcesFn(relatedResourceFhir)
-			conditionList[ndx].RelatedResourceFhir = append(conditionList[ndx].RelatedResourceFhir, relatedResourceFhir)
-		}
-
-		//sort all related resources (by date, desc)
-		conditionList[ndx].RelatedResourceFhir = utils.SortResourcePtrListByDate(conditionList[ndx].RelatedResourceFhir)
-	}
-
-	conditionList = utils.SortResourcePtrListByDate(conditionList)
-	encounterList = utils.SortResourcePtrListByDate(encounterList)
-	return conditionList, encounterList, nil
-}
-
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 // Resource Associations
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-//edges are always "strongly connected", however "source" nodes (roots, like Condition or Encounter) are only one way.
-//add an edge from every resource to its related resource. Keep in mind that FHIR resources may not contain reciprocal edges, so we ensure the graph is rooted by flipping any
-//related resources that are "Condition" or "Encounter"
-func (sr *SqliteRepository) AddReciprocalResourceAssociations(ctx context.Context, source *models.SourceCredential, resource *models.ResourceFhir, relatedSource *models.SourceCredential, relatedResource *models.ResourceFhir) error {
+func (sr *SqliteRepository) VerifyAssociationPermission(ctx context.Context, sourceUserID uuid.UUID, relatedSourceUserID uuid.UUID) error {
 	//ensure that the sources are "owned" by the same user
 	currentUser, currentUserErr := sr.GetCurrentUser(ctx)
 	if currentUserErr != nil {
 		return currentUserErr
 	}
-	if source.UserID != relatedSource.UserID {
+	if sourceUserID != relatedSourceUserID {
 		return fmt.Errorf("user id's must match when adding associations")
-	} else if source.UserID != currentUser.ID {
+	} else if sourceUserID != currentUser.ID {
 		return fmt.Errorf("user id's must match current user")
 	}
 
-	//manually create association(s) (we've tried to create using Association.Append, and it doesnt work for some reason.
-
-	if isResourceGraphSource(resource) && isResourceGraphSource(relatedResource) {
-		//handle the case where both resources are "sources" (Condition or Encounter)
-		if strings.ToLower(resource.SourceResourceType) == "condition" {
-			//condition is always the source
-			if err := sr.AddResourceAssociation(
-				ctx,
-				source,
-				resource.SourceResourceType,
-				resource.SourceResourceID,
-				source,
-				relatedResource.SourceResourceType,
-				relatedResource.SourceResourceID,
-			); err != nil {
-				//ignoring errors, could be due to duplicate edges
-				sr.Logger.Warnf("an error occurred while creating resource association: %v", err)
-			}
-		} else {
-
-			if err := sr.AddResourceAssociation(
-				ctx,
-				source,
-				relatedResource.SourceResourceType,
-				relatedResource.SourceResourceID,
-				source,
-				resource.SourceResourceType,
-				resource.SourceResourceID,
-			); err != nil {
-				//ignoring errors, could be due to duplicate edges
-				sr.Logger.Warnf("an error occurred while creating resource association: %v", err)
-			}
-		}
-
-	} else if isResourceGraphSource(resource) || isResourceGraphSink(relatedResource) {
-		//resource is a Source, or the relatedResource is a sink, create a "resource" => "relatedResource" edge
-
-		if err := sr.AddResourceAssociation(
-			ctx,
-			source,
-			resource.SourceResourceType,
-			resource.SourceResourceID,
-			source,
-			relatedResource.SourceResourceType,
-			relatedResource.SourceResourceID,
-		); err != nil {
-			//ignoring errors, could be due to duplicate edges
-			sr.Logger.Warnf("an error occurred while creating resource association: %v", err)
-		}
-
-	} else if isResourceGraphSource(relatedResource) || isResourceGraphSink(resource) {
-		//relatedResource is a Source, or the resource is a sink, create a "relatedResource" => "resource" edge
-
-		if err := sr.AddResourceAssociation(
-			ctx,
-			source,
-			relatedResource.SourceResourceType,
-			relatedResource.SourceResourceID,
-			source,
-			resource.SourceResourceType,
-			resource.SourceResourceID,
-		); err != nil {
-			//ignoring errors, could be due to duplicate edges
-			sr.Logger.Warnf("an error occurred while creating resource association: %v", err)
-		}
-
-	} else {
-		//this is a regular pair of resources, create reciprocal edges
-
-		if err := sr.AddResourceAssociation(
-			ctx,
-			source,
-			resource.SourceResourceType,
-			resource.SourceResourceID,
-			source,
-			relatedResource.SourceResourceType,
-			relatedResource.SourceResourceID,
-		); err != nil {
-			//ignoring errors, could be due to duplicate edges
-			sr.Logger.Warnf("an error occurred while creating resource association: %v", err)
-		}
-
-		if err := sr.AddResourceAssociation(
-			ctx,
-			source,
-			relatedResource.SourceResourceType,
-			relatedResource.SourceResourceID,
-			source,
-			resource.SourceResourceType,
-			resource.SourceResourceID,
-		); err != nil {
-			//ignoring errors, could be due to duplicate edges
-			sr.Logger.Warnf("an error occurred while creating resource association: %v", err)
-		}
-	}
 	return nil
 }
 
 func (sr *SqliteRepository) AddResourceAssociation(ctx context.Context, source *models.SourceCredential, resourceType string, resourceId string, relatedSource *models.SourceCredential, relatedResourceType string, relatedResourceId string) error {
+	//ensure that the sources are "owned" by the same user
+	err := sr.VerifyAssociationPermission(ctx, source.UserID, relatedSource.UserID)
+	if err != nil {
+		return err
+	}
+
 	return sr.GormClient.WithContext(ctx).Table("related_resources").Create(map[string]interface{}{
 		"resource_fhir_user_id":                      source.UserID,
 		"resource_fhir_source_id":                    source.ID,
@@ -707,15 +461,10 @@ func (sr *SqliteRepository) AddResourceAssociation(ctx context.Context, source *
 }
 
 func (sr *SqliteRepository) RemoveResourceAssociation(ctx context.Context, source *models.SourceCredential, resourceType string, resourceId string, relatedSource *models.SourceCredential, relatedResourceType string, relatedResourceId string) error {
-	currentUser, currentUserErr := sr.GetCurrentUser(ctx)
-	if currentUserErr != nil {
-		return currentUserErr
-	}
-
-	if source.UserID != relatedSource.UserID {
-		return fmt.Errorf("user id's must match when adding associations")
-	} else if source.UserID != currentUser.ID {
-		return fmt.Errorf("user id's must match current user")
+	//ensure that the sources are "owned" by the same user
+	err := sr.VerifyAssociationPermission(ctx, source.UserID, relatedSource.UserID)
+	if err != nil {
+		return err
 	}
 
 	//manually create association (we've tried to create using Association.Append, and it doesnt work for some reason.
@@ -1010,24 +759,4 @@ func sqlitePragmaString(pragmas map[string]string) string {
 		return "?" + queryStr
 	}
 	return ""
-}
-
-//source resource types are resources that are at the root of the graph, nothing may reference them directly
-func isResourceGraphSource(resource *models.ResourceFhir) bool {
-	sourceResourceTypes := []string{"condition", "encounter"}
-	return slices.Contains(sourceResourceTypes, strings.ToLower(resource.SourceResourceType)) // true
-}
-
-//sink resource types are the leaves of the graph, they must not reference anything else. (only be referenced)
-func isResourceGraphSink(resource *models.ResourceFhir) bool {
-	sinkResourceTypes := []string{"location", "binary", "device", "organization", "practitioner", "medication", "patient"}
-	return slices.Contains(sinkResourceTypes, strings.ToLower(resource.SourceResourceType)) // true
-}
-
-// helper function for GetResourceGraph, creating a "hash" for the resource
-func resourceVertexId(resource *models.ResourceFhir) string {
-	return resourceKeysVertexId(resource.SourceID.String(), resource.SourceResourceType, resource.SourceResourceID)
-}
-func resourceKeysVertexId(sourceId string, resourceType string, resourceId string) string {
-	return strings.ToLower(fmt.Sprintf("%s/%s/%s", sourceId, resourceType, resourceId))
 }
